pyproject.toml
[tool.poetry]
name = "Rag-Gallery"
version = "0.1.0"
description = ""
authors = ["Jhon Valderrama <jhonvalderramaa7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.109.1"
uvicorn = "^0.22.0"
chromadb = "^0.5.5"
onnxruntime = "^1.18.1"
pydantic-settings = "^2.4.0"
protobuf = ">=3.20.0,<3.21.0"
openai = "^1.41.0"
pymongo = "^4.8.0"
python-multipart = "^0.0.10"
shortuuid = "^1.0.13"
python-docx = "^1.1.2"
pypdf2 = "^3.0.1"
tiktoken = "^0.7.0"
numpy = "^2.1.1"
routers = "^0.10.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
